+++ chef/instrumental_tools/metadata.rb
@@ -1,0 +1,3 @@
+name "instrumental_tools"
+
+depends "packagecloud"