+++ rpm/after-remove.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+set -e
+chkconfig --del instrument_server
+exit 0